Challenge_1B/requirements.txt
pdfminer.six
sentence-transformers
transformers
torch
scikit-learn
numpy
